Fix compile

+++ bridges/pangolin-ropsten/bridge/src/command/types/affirm.rs
@@ -23,7 +23,7 @@ pub enum AffirmOpts {
         mode: AffirmMode,
         /// The block number for ethereum
         #[structopt(long)]
-        block: Option<u64>,
+        block: Option<u32>,
         /// Raw parcel json data * (not support now)
         #[structopt(long = "raw")]
         raw_json: Option<String>,
